+++ packages/stream_chat_flutter/lib/src/message_widget/bottom_row.dart
@@ -29,6 +29,7 @@ class BottomRow extends StatelessWidget {
     required this.streamChatTheme,
     required this.hasNonUrlAttachments,
     required this.streamChat,
+    this.sendingIndicatorBuilder,
     this.deletedBottomRowBuilder,
     this.onThreadTap,
     this.usernameBuilder,
@@ -39,6 +40,9 @@ class BottomRow extends StatelessWidget {
 
   /// {@macro deletedBottomRowBuilder}
   final Widget Function(BuildContext, Message)? deletedBottomRowBuilder;
+
+  /// {@macro sendingIndicatorBuilder}
+  final Widget Function(BuildContext, Message)? sendingIndicatorBuilder;
 
   /// {@macro message}
   final Message message;
@@ -147,13 +151,14 @@ class BottomRow extends StatelessWidget {
         ),
       if (showSendingIndicator)
         WidgetSpan(
-          child: SendingIndicatorWrapper(
-            messageTheme: messageTheme,
-            message: message,
-            hasNonUrlAttachments: hasNonUrlAttachments,
-            streamChat: streamChat,
-            streamChatTheme: streamChatTheme,
-          ),
+          child: sendingIndicatorBuilder?.call(context, message) ??
+              SendingIndicatorWrapper(
+                messageTheme: messageTheme,
+                message: message,
+                hasNonUrlAttachments: hasNonUrlAttachments,
+                streamChat: streamChat,
+                streamChatTheme: streamChatTheme,
+              ),
         ),
     ]);