+++ packages/stream_chat_flutter/lib/src/message_input/stream_message_input.dart
@@ -1191,7 +1191,6 @@ class StreamMessageInputState extends State<StreamMessageInput>
                 ),
               ),
               video: (attachment.file?.path ?? attachment.assetUrl)!,
-              fit: BoxFit.cover,
             ),
             Positioned(
               left: 8,