+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -1,3 +1,9 @@
+## UPCOMING
+
+✅ Added
+
+- Added `sendingIndicatorBuilder` to customize sending indicator in `StreamMessageWidget`.
+
 ## 5.0.1
 
 🔄 Changed