+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -3,6 +3,7 @@
 🔄 Changed
 
 - Updated `share_plus` dependency to `^4.5.0`
+- Expose `widthFactor` option in `MessageWidget`
 
 ## 5.0.0