+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Show message custom actions on desktop context menu.
 - Can move cursor by left/right arrow in StreamMessageInput on web/desktop.
+- Expose `widthFactor` option in `MessageWidget`
 
 ## 5.0.1