+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Upcoming
 
+✅ Added
+
+- `VideoAttachment` now uses `thumbUrl` to show the thumbnail
+  if it's available instead of generating them.
 - Expose `widthFactor` option in `MessageWidget`
 
 ## 5.0.1
@@ -555,7 +559,7 @@ typedef ActionButtonBuilder = Widget Function(
 ```
 
 > **_NOTE:_** The last parameter is the default `ActionButton`
-You can call `.copyWith` to customize just a subset of properties.
+> You can call `.copyWith` to customize just a subset of properties.
 
 - Added slow mode which allows a cooldown period after a user sends a message.
 
@@ -658,7 +662,7 @@ typedef MessageBuilder = Widget Function(
 ```
 
 > **_NOTE:_** the last parameter is the default `MessageWidget`
-You can call `.copyWith` to customize just a subset of properties
+> You can call `.copyWith` to customize just a subset of properties
 
 
 ✅ Added
@@ -728,7 +732,7 @@ typedef MessageBuilder = Widget Function(
 ```
 
 > **_NOTE:_** The last parameter is the default `MessageWidget`
-You can call `.copyWith` to customize just a subset of properties.
+> You can call `.copyWith` to customize just a subset of properties.
 
 ✅ Added